# inventory/api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Monarch Inventory API
  version: 1.0.0
  description: Complete inventory, cart, wishlist, and order API for Monarch.
servers:
  - url: http://localhost:8082
    description: Inventory service (direct)
  - url: http://localhost:8080/inventory
    description: Via API gateway
...
paths:
...
  /cart:
    get:
      tags: [Cart]
      summary: Get customer cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart returned
    post:
      tags: [Cart]
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Cart updated
    delete:
      tags: [Cart]
      summary: Clear cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart cleared
...
  /wishlist:
    get:
      tags: [Wishlist]
      summary: Get customer wishlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wishlist returned
    post:
      tags: [Wishlist]
      summary: Add item to wishlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToWishlistRequest"
      responses:
        "200":
          description: Wishlist updated
    delete:
      tags: [Wishlist]
      summary: Clear wishlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wishlist cleared
...
  /orders:
    post:
      tags: [Orders]
      summary: Create order from cart
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Order created
    get:
      tags: [Orders]
      summary: Get customer orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Orders returned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AddToCartRequest:
      type: object
      required: [itemId, quantity]
      properties:
        itemId:
          type: string
        quantity:
          type: integer
          minimum: 1
...
    AddToWishlistRequest:
      type: object
      required: [itemId]
      properties:
        itemId:
          type: string
...